refactor(agent): extract anthropic-compat helpers into aux_anthropic_compat (auxiliary_client.py R4-S1, epic #78647)
